introduce formal rule visitor
